Move currency related tests to separate mod
